Don't include groups in the content of groups
